refactor: split hole/[number].tsx into hooks + components
